requirements.txt
numpy
pandas
matplotlib
opencv-python
pillow
pandas
scikit-learn
tk